requirements.txt
#rodar um pip install -r requirements.txt para ter as dependências necessárias
Pillow==8.3.1
tk
